Implement the standard algorithm
This algorithm works by reading N/2 amount of data at a time into a
buffer of size N. If the pattern is in the buffer, we find the closest
newline byte and stop printing there. Otherwise, we print the first half
of the buffer and shift everything up.

The advantage of this algorithm is that it works efficiently on very
large files. We only use N bytes of buffer regardless of the filesize.

The disadvantage is that it might not be correct for files with lines
longer than N/2.