Major refactor to use traits and a configuration struct
This will make adding flags and changing behavior very easy. The trait
handles finding the patterns and will call the relevant methods for
before / at / after the match.

One downside is that line mode *does* read line by line, which requires a
variable amount of memory and patterns with newlines cannot be used. The
upside is that the algorithm cannot fail to find the previous line before
a match.